refactor: refactor all grpahics resource desc
