+++ src/Rendering/GraphicsResourcePrefix.h
@@ -11,6 +11,40 @@ enum class MemoryUsage
     Count
 };
 
+enum class LogicOp : uint8_t
+{
+    Clear,
+    And,
+    AndReverse,
+    Copy,
+    AndInverted,
+    NoOp,
+    Xor,
+    Or,
+    Nor,
+    Equivalent,
+    Invert,
+    OrReverse,
+    CopyInverted,
+    OrInverted,
+    Nand,
+    Set,
+    Count
+};
+
+enum class CompareOp : uint8_t
+{
+    Never,
+    Less,
+    Equal,
+    LessEqual,
+    Greater,
+    NotEqual,
+    GreaterEqual,
+    AlwaysTrue,
+    Count
+};
+
 enum class BufferUsage
 {
     Vertex,
@@ -22,57 +56,6 @@ enum class BufferUsage
     Indirect,
     RayTracing,
     AccelerationStructure,
-    Count
-};
-
-enum TextureUsageBits : uint8_t
-{
-    Sampled      = 1 << 0, // SRV
-    Storage      = 1 << 1, // UAV
-    RenderTarget = 1 << 2, // RTV
-    DepthStencil = 1 << 3, // DSV
-    Count
-};
-
-inline TextureUsageBits operator|(TextureUsageBits a, TextureUsageBits b)
-{
-    return static_cast<TextureUsageBits>(static_cast<uint8_t>(a) | static_cast<uint8_t>(b));
-}
-
-inline TextureUsageBits operator&(TextureUsageBits a, TextureUsageBits b)
-{
-    return static_cast<TextureUsageBits>(static_cast<uint8_t>(a) & static_cast<uint8_t>(b));
-}
-
-enum class TextureType : uint8_t
-{
-    Tex1D,
-    Tex2D,
-    Tex3D,
-    Count
-};
-
-enum class SamplerFilter
-{
-    Nearest,
-    Linear,
-    Count
-};
-
-enum class SamplerAddressMode
-{
-    Repeat,
-    MirroredRepeat,
-    ClampToEdge,
-    ClampToBorder,
-    MirrorClampToEdge,
-    Count
-};
-
-enum class SamplerMipmapMode
-{
-    Nearest,
-    Linear,
     Count
 };
